test: consolidate virtual app tests
Merge provider API, offline resource and publication confirmation tests into terminal/tests.py, preserving all 82 test cases.

Initialize the asset automation cache for test isolation.

Validation: all 85 related tests passed.
